// src/server/app/enum/profil.enum.ts
export enum ProfilEnum {
  COORDINATEUR_REGION = 'COORDINATEUR_REGION', // Coordinateur PILOTE régional
  COORDINATEUR_DEPARTEMENT = 'COORDINATEUR_DEPARTEMENT', // Coordinateur PILOTE départemental
  CABINET_MTFP = 'CABINET_MTFP', // Cabinet MTFP
  PM_ET_CABINET = 'PM_ET_CABINET', // Première Ministre et cabinet
  PR = 'PR', // Présidence de la République
  CABINET_MINISTERIEL = 'CABINET_MINISTERIEL', // Cabinets ministériels
  DIR_ADMIN_CENTRALE = 'DIR_ADMIN_CENTRALE', // Direction d'administration centrale
  DROM = 'DROM', // DROM/Outre-Mer
  PREFET_DEPARTEMENT = 'PREFET_DEPARTEMENT', // Préfet de département et collaborateurs
  PREFET_REGION = 'PREFET_REGION', // Préfet de région et collaborateurs
  RESPONSABLE_DEPARTEMENT = 'RESPONSABLE_DEPARTEMENT', // Responsable local départemental
  RESPONSABLE_REGION = 'RESPONSABLE_REGION', // Responsable local régional
  SERVICES_DECONCENTRES_DEPARTEMENT = 'SERVICES_DECONCENTRES_DEPARTEMENT', // Services déconcentrés départementaux
  SERVICES_DECONCENTRES_REGION = 'SERVICES_DECONCENTRES_REGION', // Services déconcentrés régionaux
  SECRETARIAT_GENERAL = 'SECRETARIAT_GENERAL', // Secrétariat général de ministère (cormod)
  DIR_PROJET = 'DIR_PROJET', // Directeur de projet
  EQUIPE_DIR_PROJET = 'EQUIPE_DIR_PROJET', // Équipe de Directeur de projet
  DITP_ADMIN = 'DITP_ADMIN', // DITP - Admin
  DITP_PILOTAGE = 'DITP_PILOTAGE', // DITP - Pilotage
}